+++ examples/minimal-with-gulp/src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 // Material-UI
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
-import Tooltip from '@material-ui/core/Tooltip';
+import MuiTooltip from '@material-ui/core/Tooltip';
 import IconButton from '@material-ui/core/IconButton';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
@@ -12,10 +12,9 @@ import { withStyles } from '@material-ui/core/styles';
 import DayIcon from '@material-ui/icons/WbSunnyOutlined';
 import NightIcon from '@material-ui/icons/Brightness3Outlined';
 
-import 'typeface-roboto';
-
 // Lattice
 import { Widget } from '@latticejs/widgets';
+import { PieChart, Pie, Tooltip } from '@latticejs/mui-recharts';
 
 // Custom Style
 const styles = theme => ({
@@ -39,10 +38,43 @@ const styles = theme => ({
   }
 });
 
+const Chart = props => (
+  <PieChart width={400} height={400}>
+    <Pie data={props.deps} dataKey="value" cx={200} cy={200} outerRadius={60} fill="#8884d8" />
+    <Pie data={props.devDeps} dataKey="value" cx={200} cy={200} innerRadius={70} outerRadius={90} fill="#82ca9d" />
+    <Tooltip />
+  </PieChart>
+);
+
 class App extends Component {
   handleNightModeChange = () => {
     const { updateTheme, nightMode } = this.props;
     updateTheme(!nightMode);
+  };
+
+  state = {
+    data: {
+      deps: [
+        { name: '@latticejs/widgets', value: 123369 },
+        { name: '@material-ui/core', value: 392567 },
+        { name: '@material-ui/icons', value: 3630161 },
+        { name: 'react', value: 5529 },
+        { name: 'react-dom', value: 92455 },
+        { name: 'typeface-roboto', value: 3064 }
+      ],
+      devDeps: [
+        { name: 'babel-core', value: 539213 },
+        { name: 'babel-loader', value: 46534 },
+        { name: 'babel-plugin-transform-class-properties', value: 371741 },
+        { name: 'babel-preset-env', value: 61171 },
+        { name: 'babel-preset-react', value: 129792 },
+        { name: 'css-loader', value: 214658 },
+        { name: 'file-loader', value: 179134 },
+        { name: 'html-webpack-plugin', value: 625851 },
+        { name: 'style-loader', value: 170557 },
+        { name: 'webpack', value: 1900109 }
+      ]
+    }
   };
 
   render() {
@@ -53,41 +85,38 @@ class App extends Component {
         <AppBar position="static" className={classes.appBar}>
           <Toolbar>
             <Typography variant="h6" color="inherit" className={classes.flex}>
-              Basic Example
+              Minimal With Gulp Example
             </Typography>
-            <Tooltip title="Toggle Night Mode" enterDelay={300}>
+            <MuiTooltip title="Toggle Night Mode" enterDelay={300}>
               <IconButton onClick={this.handleNightModeChange} color="inherit">
                 {nightMode ? <DayIcon /> : <NightIcon />}
               </IconButton>
-            </Tooltip>
+            </MuiTooltip>
           </Toolbar>
         </AppBar>
         <Grid container>
           <Grid item xs={12}>
-            <Grid container justify="space-around" spacing={Number('0')}>
+            <Grid container justify="space-around" spacing={Number(0)}>
               <Grid item>
-                <Widget className={classes.widget} title="Introduction" border="bottom">
-                  <Typography variant="subtitle1">Welcome to Lattice</Typography>
+                <Widget className={classes.widget} title="About" border="bottom">
+                  <Typography variant="body1">
+                    This is a minimal demo with respect to:
+                    <br />
+                    - minimal tooling (webpack and babel),
+                    <br />
+                    - no react-scripts,
+                    <br />- consuming Lattice packages directly.
+                  </Typography>
                 </Widget>
               </Grid>
               <Grid item>
-                <Widget className={classes.widget} title="Material" border="bottom">
-                  <Typography variant="subtitle1">Material UI integration</Typography>
-                </Widget>
-              </Grid>
-              <Grid item>
-                <Widget className={classes.widget} title="Recharts" border="bottom">
-                  <Typography variant="subtitle1">with Material style</Typography>
-                </Widget>
-              </Grid>
-              <Grid item>
-                <Widget className={classes.widget} title="D3" border="bottom">
-                  <Typography variant="subtitle1">React + D3 integration</Typography>
-                </Widget>
-              </Grid>
-              <Grid item>
-                <Widget className={classes.widget} title="React Virtualized" border="bottom">
-                  <Typography variant="subtitle1">Infinite list support</Typography>
+                <Widget className={classes.widget} title="Bundle data" border="bottom">
+                  <div>
+                    {Chart(this.state.data)}
+                    <Typography variant="caption" align="center">
+                      NOTE: Actual sizes might be smaller if only parts of the package are used.
+                    </Typography>
+                  </div>
                 </Widget>
               </Grid>
             </Grid>